+++ apps/admin/src/components/Navbar.tsx
@@ -43,7 +43,7 @@ const linkMap = {
   Shifts: "/shifts",
   "Event Check-in": "/event-checkin",
   Merch: "/merch",
-  Massmailer: "/email-maker",
+  Massmailer: "/massmailer",
   Attendance: "/attendance-view"
 };
