+++ apps/admin/src/routes.tsx
@@ -7,7 +7,7 @@ import Sponsors from "./routes/pages/Sponsors";
 import Merch from "./routes/pages/Merch";
 import EventCheckin from "./routes/pages/EventCheckin";
 import Attendance from "./routes/pages/Attendance";
-import EmailMaker from "./routes/pages/EmailMaker";
+import Massmailer from "./routes/pages/Massmailer";
 import Speakers from "./routes/pages/Speakers";
 import Shifts from "./routes/pages/Shifts";
 
@@ -20,7 +20,7 @@ const routes = [
   { path: "/sponsors", element: <Sponsors /> },
   { path: "/merch", element: <Merch /> },
   { path: "/event-checkin", element: <EventCheckin /> },
-  { path: "/email-maker", element: <EmailMaker /> },
+  { path: "/massmailer", element: <Massmailer /> },
   { path: "/attendance-view", element: <Attendance /> },
   { path: "/speakers", element: <Speakers /> },
   { path: "/shifts", element: <Shifts /> }
